move ElementInspector and related components into separate dir
