+++ extension/src/frontend/view/ElementInspector/ElementPopover.tsx
