+++ extension/src/frontend/view/ElementInspector/assertions/VisibilityAssertionForm.tsx
