+++ extension/src/frontend/view/ElementInspector/index.ts
@@ -1,0 +1,1 @@
+export * from './ElementInspector'
